requirements.txt
fastapi
uvicorn
google-cloud-firestore
pytest